# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Cool API",
        "description": "An API of cool stuffs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/mobils": {
            "get": {
                "tags": [
                    "Mobil"
                ],
                "summary": "Get list of Mobil",
                "operationId": "411966e9a27c8d0a45af5d95616c1986",
                "parameters": [
                    {
                        "name": "merek",
                        "in": "query",
                        "description": "Filter by merek",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Filter by model",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ketersediaan",
                        "in": "query",
                        "description": "Filter by ketersediaan (availability)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Mobil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Mobil Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "merek": {
                                                        "type": "string",
                                                        "example": "Toyota"
                                                    },
                                                    "model": {
                                                        "type": "string",
                                                        "example": "Camry"
                                                    },
                                                    "nomor_plat": {
                                                        "type": "string",
                                                        "example": "ABC1234"
                                                    },
                                                    "tarif_sewa_per_hari": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 500000
                                                    },
                                                    "ketersediaan": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mobil"
                ],
                "summary": "Store a new Mobil",
                "operationId": "6d2ef423591bc77d14b846974eeb6f5e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "merek": {
                                                "type": "string"
                                            },
                                            "model": {
                                                "type": "string"
                                            },
                                            "nomor_plat": {
                                                "type": "string"
                                            },
                                            "tarif_sewa_per_hari": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "ketersediaan": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "merek": "Toyota",
                                    "model": "Camry",
                                    "nomor_plat": "ABC1234",
                                    "tarif_sewa_per_hari": 500000,
                                    "ketersediaan": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mobil created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Mobil Berhasil Dibuat."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "merek": {
                                                    "type": "string",
                                                    "example": "Toyota"
                                                },
                                                "model": {
                                                    "type": "string",
                                                    "example": "Camry"
                                                },
                                                "nomor_plat": {
                                                    "type": "string",
                                                    "example": "ABC1234"
                                                },
                                                "tarif_sewa_per_hari": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500000
                                                },
                                                "ketersediaan": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        },
                                        "errors": {
                                            "properties": {
                                                "merek": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The merek field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "model": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The model field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "nomor_plat": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The nomor plat field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "tarif_sewa_per_hari": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The tarif sewa per hari field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "ketersediaan": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The ketersediaan field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}